+++ packages/@overeng/genie/nix/build.nix
@@ -19,7 +19,7 @@ let
     # to download binaries for all platforms (linux/darwin x x64/arm64), so the hash is
     # the same regardless of where the build runs.
     pnpmDepsHash = "sha256-RVbdCRnWYxoDip1R3AiPuLm/CF7ZIqotNAKcLyAvkbs=";
-    lockfileHash = "sha256-jYIRxXlmlMW305bGUYJyzcVqf/ZaaBVrYDqdCeMBQUc=";
+    lockfileHash = "sha256-UdDpyWjp8SdE9BsMhHWlbvYpQ3fOo1sMShQKXF4Tkak=";
     packageJsonDepsHash = "sha256-vGC5f0D1JP3TyazuWez475pUTd4XD/slhT89ybmI0Cs=";
     inherit gitRev commitTs dirty;
   };